Pie Puller Spring-Security-Login-And-Registration
Current functions:
User may register a unique account and must fill out the entire form. 
If user fails to sign up with a unique username then they will be returned to the registration form again. 

After a user has registered, they will be automatically assigned the role of User and have access to the homepage. Soon to add Edit Profile page for making changes to some user details.

On the homepage:
User may logout, if user is also an ADMIN or a MANAGER then they will have access to links directing to a ADMIN or MANAGER only resource.