Employee Registration view files
Adjusted files so that they referred to the employee database, resulting in all registrations from the employee-registration-form to go to the employee table in the database rather than the customer table. 

employee-registration-confirmation.jsp was changed to note to the admin/manager adding employee to write down user login credentials to present to the new employee.