CustomerService Interface, CustomerServiceImpl
CustomerServiceImpl implements most of the methods that were in the UserService previously. Due to the way that the customer authentication config file attempts to authenticate users, the CustomerServiceImpl attempts to find the user in the database. If the user returns as null, then it will query the employee table for the user and if it does find a user, it logs that user in. If not, then it will return Invalid username or password.